// src/modules/preventive-maintenance-documents/models/folders.enum.ts
export enum Folders {
  PM_INSTRUCTIONS = 'PM_INSTRUCTIONS',
  COMMERCIAL_DOCS = 'COMMERCIAL_DOCS',
  PM_REPORTS = 'PM_REPORTS',
  IMAGES = 'IMAGES',
  VIDEOS = 'VIDEOS',
  DOCUMENT_UPLOAD = 'DOCUMENT_UPLOAD',
  ACTIVITY = 'ACTIVITY',
  PROCEDURES = 'PROCEDURES',
  SUBMIT_FOR_REVIEW = 'SUBMIT_FOR_REVIEW',
}